Avances en ventas persistentes
git-svn-id: http://openomoikane2.googlecode.com/svn/trunk/Omoikane@292 a36d047a-ba86-11dd-b55b-2da8d274f894

+++ main/src/omoikane/sistema/Comprobantes.groovy
@@ -40,7 +40,7 @@ import groovy.inspect.swingui.*
     /**
      * Método que utiliza hibernate para acceder a los datos de la venta en lugar de nadesico (versión antigua ticket())
      */
-    public void ticketVenta( Integer idVenta ) {
+    public void ticketVenta( Long idVenta ) {
         LegacyVenta venta = entityManager.find(LegacyVenta.class, idVenta);
         data              = venta.properties;
         data.date         = data.fechaHora
